Fix all 9 ESLint errors to unblock CI pipeline
- usePretext.js: Wrap fallback measure() calls in requestAnimationFrame to
  avoid synchronous setState in effect body (react-hooks/set-state-in-effect)
- OwnerDashboard.jsx: Move setContractOwner('') into microtask via
  Promise.resolve() to avoid synchronous setState in effect
- RichText.jsx: Extract CHIP_STYLES, RICH_TEXT_STYLES, parseRichSpecs to
  dedicated richTextStyles.js (react-refresh/only-export-components)
- Dashboard.jsx: Import CHIP_STYLES from new richTextStyles module
- pretext.js: Export graphemeSegmenter() function (was defined but unused)

All 698 tests pass (489 contract + 174 SDK + 35 frontend). Lint: 0 errors.
