+++ nexus-app/src/components/richTextStyles.js
@@ -1,0 +1,81 @@
+/**
+ * Shared chip/badge styles and markup parser for RichText.
+ * Extracted from RichText.jsx so that file only exports React components
+ * (required by react-refresh/only-export-components).
+ */
+
+import { FONTS } from '../config/designTokens.js';
+
+// ---------------------------------------------------------------------------
+// Pre-built chip/badge styles matching NEXUS design language
+// ---------------------------------------------------------------------------
+
+const chipBase = 'inline-flex items-center px-2.5 min-h-[22px] rounded-full text-xs font-bold whitespace-nowrap border';
+
+export const CHIP_STYLES = {
+  status:   `${chipBase} bg-nexus-cyan/10 text-nexus-cyan border-nexus-cyan/20`,
+  success:  `${chipBase} bg-nexus-green/10 text-nexus-green border-nexus-green/20`,
+  warning:  `${chipBase} bg-nexus-amber/10 text-nexus-amber border-nexus-amber/20`,
+  danger:   `${chipBase} bg-nexus-red/10 text-nexus-red border-nexus-red/20`,
+  info:     `${chipBase} bg-nexus-purple/10 text-nexus-purple border-nexus-purple/20`,
+  neutral:  `${chipBase} bg-white/5 text-nexus-text-dim border-nexus-border`,
+  mention:  `${chipBase} bg-nexus-cyan/15 text-nexus-cyan border-nexus-cyan/25`,
+};
+
+export const RICH_TEXT_STYLES = {
+  code: 'px-1.5 py-0.5 rounded-md bg-white/8 font-mono text-xs text-nexus-text',
+  link: 'text-nexus-cyan underline underline-offset-2 decoration-1',
+  bold: 'font-semibold text-nexus-text',
+  dim: 'text-nexus-text-dim',
+};
+
+// ---------------------------------------------------------------------------
+// Helper: build specs from a simple markup DSL
+// ---------------------------------------------------------------------------
+
+/**
+ * Parse simple inline markup into specs array.
+ * Supports: `code`, **bold**, {chip:tone:label}
+ *
+ * Example:
+ *   parseRichSpecs("Ship `pre-wrap` for {chip:status:blocked} by **vertical text**")
+ *
+ * @param {string} markup
+ * @param {{ bodyFont?: string, codeFont?: string, chipFont?: string }} [fonts]
+ * @returns {Array<import('../lib/pretext.js').InlineSpec|import('../lib/pretext.js').ChipSpec>}
+ */
+export function parseRichSpecs(markup, fonts) {
+  const bodyFont = fonts?.bodyFont ?? FONTS.body;
+  const codeFont = fonts?.codeFont ?? FONTS.mono;
+  const chipFont = fonts?.chipFont ?? FONTS.chip;
+
+  const specs = [];
+  // Regex: `code`, **bold**, {chip:tone:label}, or plain text
+  // The final alternative matches any single special char that didn't form
+  // a complete token, so lone { * ` are preserved as plain text.
+  const pattern = /`([^`]+)`|\*\*([^*]+)\*\*|\{chip:(\w+):([^}]+)\}|([^`*{]+)|(.)/g;
+  let match;
+
+  while ((match = pattern.exec(markup)) !== null) {
+    if (match[1] !== undefined) {
+      // `code`
+      specs.push({ kind: 'text', text: match[1], font: codeFont, className: RICH_TEXT_STYLES.code, chromeWidth: 12 });
+    } else if (match[2] !== undefined) {
+      // **bold**
+      specs.push({ kind: 'text', text: match[2], font: bodyFont, className: RICH_TEXT_STYLES.bold });
+    } else if (match[3] !== undefined && match[4] !== undefined) {
+      // {chip:tone:label}
+      const tone = match[3];
+      const label = match[4];
+      specs.push({ kind: 'chip', label, font: chipFont, className: CHIP_STYLES[tone] ?? CHIP_STYLES.neutral, chromeWidth: 22 });
+    } else if (match[5] !== undefined) {
+      // plain text
+      specs.push({ kind: 'text', text: match[5], font: bodyFont });
+    } else if (match[6] !== undefined) {
+      // single unmatched special char — preserve as plain text
+      specs.push({ kind: 'text', text: match[6], font: bodyFont });
+    }
+  }
+
+  return specs;
+}
